# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Capyba",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista os posts do usuário autenticado",
                "description": "Retorna uma lista de posts do usuário autenticado com base nos parâmetros de busca.",
                "operationId": "5f292ed47bdbfa79356750be0807450c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtro de busca por título ou conteúdo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Quantidade de posts por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página a ser retornada",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "ordering",
                        "in": "query",
                        "description": "Campo de ordenação (exemplo: '-id' para decrescente ou 'id' para crescente)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "-id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum post encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhum post encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/store": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cria um novo post com upload de imagem",
                "description": "Cria um novo post e faz upload de uma imagem, utilizando multipart/form-data.",
                "operationId": "c7329c8eff6a626367880153b14203be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Título do post",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Conteúdo do post",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Imagem do post",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o Post"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/show/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe o post selecionado",
                "description": "Retorna o post selecionado.",
                "operationId": "c426c7168bc5e04e27d5192c1b3c4604",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post a ser visualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum post encontrado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/update/{id}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualiza um post existente",
                "description": "Atualiza um post e permite o upload de uma nova imagem. Apenas usuários com e-mail verificado podem realizar essa ação.",
                "operationId": "a0183fc286415272ef49e7bba3b58703",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do post",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Conteúdo do post",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Nova imagem do post",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "_method": {
                                        "description": "Método PUT (method spoofing)",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário não verificado"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    },
                    "500": {
                        "description": "Erro ao atualizar o Post"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/delete/{id}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Deleta um post existente",
                "description": "Deleta um post. Apenas usuários com e-mail verificado podem realizar essa ação.",
                "operationId": "f6db0f869393793cfa6139294448a1c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deletado com sucesso"
                    },
                    "403": {
                        "description": "Usuário não verificado"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    },
                    "500": {
                        "description": "Erro ao deletar o Post"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Post": {
                "title": "Post",
                "required": [
                    "id",
                    "title",
                    "content",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use o token fornecido pelo Sanctum. Exemplo: 'Bearer {token}'",
                "scheme": "bearer"
            }
        }
    },
...
}